backend/requirements.txt
fastapi
uvicorn[standard]
PyJWT
python-multipart

